+++ src/tokenise.h
@@ -2,6 +2,7 @@
 #define TOKENISE_H
 
 #include <stdbool.h>
+#include <stdint.h>
 
 typedef struct Token Token;
 
@@ -15,7 +16,7 @@ bool consumeFor(void);
 bool consumeSizeof(void);
 Token *consumeStrLit(void);
 void expect(char *op);
-int expectNumber(void);
+int64_t expectNumber(void);
 Token *expectIdent(void);
 void tokenise(const char *file_path);
 bool isEOF(void);
